requirements.txt
streamlit
pyocr
Pillow
PyPDF2
PyMuPDF